+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Require `click` >=8.2 ([#869])
+- Replace `mattermostdriver` by `mattermostautodriver` ([#988])
+
+  The Mattermost URL must now be configured **without** `/api/v4`. For example:
+
+  ```diff
+  - http://127.0.0.1:8065/api/v4
+  + http://127.0.0.1:8065
+  ```
 
 ### Removed
 - Drop support for Python 3.9 ([#887])
@@ -92,6 +100,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.2.0]: https://github.com/maxbrunet/mmemoji/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/maxbrunet/mmemoji/releases/tag/v0.1.0
 
+[#988]: https://github.com/maxbrunet/mmemoji/issues/988
 [#887]: https://github.com/maxbrunet/mmemoji/issues/887
 [#709]: https://github.com/maxbrunet/mmemoji/issues/709
 [#604]: https://github.com/maxbrunet/mmemoji/issues/604